+++ NetBeansProjects/bon-hiato2/src/po/ProductOwner.java
@@ -28,8 +28,14 @@ public class ProductOwner extends javax.swing.JFrame {
      */
     BD bd = new BD();
     Connection connection = bd.conexion();
+    private String develop;
     public ProductOwner() {
         initComponents();
+    }
+    public ProductOwner(String deve){
+        initComponents();
+        develop=deve;
+        jLabel1.setText("<Html>Logueado como: <br>"+develop+"</Html>");
         try{
             Statement st = connection.createStatement();
             ResultSet rs = st.executeQuery("select * from projects");
@@ -56,6 +62,7 @@ public class ProductOwner extends javax.swing.JFrame {
         selectProductOwner = new javax.swing.JButton();
         newProductOwner = new javax.swing.JButton();
         remProductOwner = new javax.swing.JButton();
+        jLabel1 = new javax.swing.JLabel();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setTitle("Bon-Hiato");
@@ -90,29 +97,37 @@ public class ProductOwner extends javax.swing.JFrame {
             }
         });
 
+        jLabel1.setText("jLabel1");
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addGap(0, 101, Short.MAX_VALUE)
+                .addContainerGap()
+                .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 258, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 46, Short.MAX_VALUE)
+                .addComponent(logoutProductOwner)
+                .addContainerGap())
+            .addGroup(layout.createSequentialGroup()
+                .addGap(103, 103, 103)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
                     .addComponent(remProductOwner, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                     .addComponent(newProductOwner, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                     .addComponent(selectProductOwner, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                     .addComponent(projectProductOwner, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(productOwnerSelector, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addGap(60, 60, 60)
-                .addComponent(logoutProductOwner)
-                .addContainerGap())
+                    .addComponent(productOwnerSelector))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
                 .addContainerGap()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(logoutProductOwner)
-                    .addComponent(productOwnerSelector))
+                    .addComponent(jLabel1))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 41, Short.MAX_VALUE)
+                .addComponent(productOwnerSelector)
                 .addGap(30, 30, 30)
                 .addComponent(projectProductOwner, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(29, 29, 29)
@@ -121,7 +136,7 @@ public class ProductOwner extends javax.swing.JFrame {
                 .addComponent(newProductOwner)
                 .addGap(18, 18, 18)
                 .addComponent(remProductOwner)
-                .addContainerGap(70, Short.MAX_VALUE))
+                .addContainerGap())
         );
 
         pack();
@@ -130,8 +145,7 @@ public class ProductOwner extends javax.swing.JFrame {
     private void selectProductOwnerActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_selectProductOwnerActionPerformed
         // TODO add your handling code here:
         String choice=projectProductOwner.getSelectedItem();
-        Developer developer = new Developer();
-        ProductBacklog productBacklog = new ProductBacklog(choice); 
+        ProductBacklog productBacklog = new ProductBacklog(choice,develop); 
         productBacklog.setLocationRelativeTo(null);
         productBacklog.setVisible(true);
         dispose();
@@ -139,7 +153,7 @@ public class ProductOwner extends javax.swing.JFrame {
 
     private void newProductOwnerActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_newProductOwnerActionPerformed
         // TODO add your handling code here:
-        CreateProject createProject = new CreateProject();
+        CreateProject createProject = new CreateProject(develop);
         createProject.setLocationRelativeTo(null);
         createProject.setVisible(true);
         dispose();
@@ -155,19 +169,22 @@ public class ProductOwner extends javax.swing.JFrame {
 
     private void remProductOwnerActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_remProductOwnerActionPerformed
         // TODO add your handling code here:
-        try{
-            String query = " DELETE FROM projects WHERE name = ?";
-            
-            PreparedStatement preparedStmt = connection.prepareStatement(query);
-            preparedStmt.setString(1, projectProductOwner.getSelectedItem());
-            preparedStmt.execute();
-        }catch(SQLException | HeadlessException e){
-            JOptionPane.showMessageDialog(null, e);
+        int respuesta=JOptionPane.showConfirmDialog(null, "¿Estas seguro que quieres eliminar esta tarea?");
+        if(respuesta==0){
+            try{
+                String query = " DELETE FROM projects WHERE name = ?";
+
+                PreparedStatement preparedStmt = connection.prepareStatement(query);
+                preparedStmt.setString(1, projectProductOwner.getSelectedItem());
+                preparedStmt.execute();
+            }catch(SQLException | HeadlessException e){
+                JOptionPane.showMessageDialog(null, e);
+            }
+            ProductOwner productOwner = new ProductOwner(develop);
+            productOwner.setLocationRelativeTo(null);
+            productOwner.setVisible(true);
+            dispose();
         }
-        ProductOwner productOwner = new ProductOwner();
-        productOwner.setLocationRelativeTo(null);
-        productOwner.setVisible(true);
-        dispose();
     }//GEN-LAST:event_remProductOwnerActionPerformed
 
     /**
@@ -206,6 +223,7 @@ public class ProductOwner extends javax.swing.JFrame {
     }
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JLabel jLabel1;
     private javax.swing.JButton logoutProductOwner;
     private javax.swing.JButton newProductOwner;
     private javax.swing.JLabel productOwnerSelector;
